LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/parsing/column-width-computed-expected.txt


PASS Property column-width value 'auto' 
PASS Property column-width value 'calc(10px + 0.5em)' 
FAIL Property column-width value 'calc(10px - 0.5em)' assert_equals: computed value should round-trip expected "0px" but got "auto"
